Switch Time::elapsedMsecs to return qint64.
When dealing with elapsed times you almost always want signed integers
because zero crossings can cause crazy large numbers.

For unsigned integers: (small number - small number) => GIGANTIC NUMBER
For signed integers: (small number - small number) = +/- small number

+++ src/controllers/softtakeover.cpp
@@ -79,7 +79,7 @@ bool SoftTakeover::ignore(ControlObject* control, double newParameter) {
     //      of the current value of the control
     //  - it's been awhile since the controller last affected this control
 
-    uint currentTime = Time::elapsedMsecs();
+    qint64 currentTime = Time::elapsedMsecs();
     // We will get a sudden jump if we don't ignore the first value.
     if (m_time == 0) {
         ignore = true;
