+++ src/controllers/softtakeover.h
@@ -31,9 +31,9 @@ class SoftTakeover {
     // regardless. This allows quickly whipping controls to work while retaining
     // the benefits of soft-takeover for slower movements.  Setting this too
     // high will defeat the purpose of soft-takeover.
-    static const uint SUBSEQUENT_VALUE_OVERRIDE_TIME_MILLIS = 50;
+    static const qint64 SUBSEQUENT_VALUE_OVERRIDE_TIME_MILLIS = 50;
 
-    uint m_time;
+    qint64 m_time;
     double m_prevParameter;
     double m_dThreshold;
 };
